Cast operand to long to prevent any possible int overflow

+++ src/main/java/uk/co/bconline/ndelius/service/impl/OIDUserDetailsService.java
@@ -137,7 +137,7 @@ public class OIDUserDetailsService implements OIDUserService, UserDetailsService
 						.countLimit(pageSize * page)
 						.filter(filter))
 				.spliterator(), false)
-				.skip(pageSize * (page-1))
+				.skip((long) pageSize * (page-1))
 				.collect(toList());
 	}
